Refactor remote config tests and enhance application launch process
- Updated remote config test to check for individual config values using `find.textContaining`.
- Modified `launchApplication` to use `pumpWidget` followed by `pumpAndTrySettle` for better stability.
- Improved permission handling in `ensureAuthenticated` with additional wait time for dialog dismissal.
- Enhanced topic selection process with retry logic for tapping the search button.
- Added new files for analytics UI components, including charts and heatmaps for better data visualization.
- Introduced `AppStateView`, `ChartEmptyState`, `DonutBreakdownChart`, `YearBarChart`, and `YearHeatmap` for consistent presentation of empty and data-driven states.
- Implemented formatting functions for chart values to improve readability.
- Created responsive metric grid layout for analytics display.

+++ lib/data/models/dashboard_report_data.dart
@@ -4,7 +4,7 @@ import 'ranked_entity.dart';
 class DashboardReportData {
   final String topic;
   final int totalPublications;
-  final int? averageCitations;
+  final double? averageCitations;
   final int? mostActiveYear;
   final String? topJournal;
   final String? topAuthor;
